+++ TelegramBots_V4/Commands/MenuCommands.cs
@@ -12,14 +12,11 @@ namespace TelegramBots_V4.Commands
     [Action("/menu", "меню")]
     public async void Menu()
     {
-      if (Context.Update.CallbackQuery is not null)
+      if (Context.Update.CallbackQuery is not null && Context.Update.CallbackQuery.Message is not null)
       {
-        if (Context.Update.CallbackQuery.Message is not null)
-        {
-          await Context.Bot.Client.DeleteMessageAsync(Context.Update.CallbackQuery.Message.Chat.Id, Context.Update.CallbackQuery.Message.MessageId);
-          await Context.Bot.Client.AnswerCallbackQueryAsync(Context.Update.CallbackQuery.Id);
-          await Context.Bot.Client.SendTextMessageAsync(Context.Update.CallbackQuery.Message.Chat.Id, "все команды", replyMarkup: FormInlineCommands.AllCommands());
-        }
+        await Context.Bot.Client.DeleteMessageAsync(Context.Update.CallbackQuery.Message.Chat.Id, Context.Update.CallbackQuery.Message.MessageId);
+        await Context.Bot.Client.AnswerCallbackQueryAsync(Context.Update.CallbackQuery.Id);
+        await Context.Bot.Client.SendTextMessageAsync(Context.Update.CallbackQuery.Message.Chat.Id, "все команды", replyMarkup: FormInlineCommands.AllCommands());
       }
       else
       {
